requirements-test.txt
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0 